02_Kafka/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --python-version 3.11 requirements.in -o requirements.txt
aiokafka==0.12.0
    # via -r requirements.in
async-timeout==5.0.1
    # via aiokafka
kafka-python==2.3.0
    # via -r requirements.in
packaging==25.0
    # via aiokafka
typing-extensions==4.15.0
    # via aiokafka
